Move refresh method to object sets and fix operator to refresh colours
* Move refresh object sets colours method to object_sets file
* Fix bad reference to function call refresh object sets
